Rust 1.49 Clippy Fixes

+++ crates/nu_plugin_inc/src/inc.rs
@@ -79,7 +79,7 @@ impl Inc {
                 Ok(UntaggedValue::int(i + 1).into_value(value.tag()))
             }
             UntaggedValue::Primitive(Primitive::Filesize(b)) => {
-                Ok(UntaggedValue::filesize(b + 1 as u64).into_value(value.tag()))
+                Ok(UntaggedValue::filesize(b + 1_u64).into_value(value.tag()))
             }
             UntaggedValue::Primitive(Primitive::String(ref s)) => {
                 Ok(self.apply(&s)?.into_value(value.tag()))
